+++ eslint.config.ts
@@ -1,19 +1,32 @@
 import js from "@eslint/js";
 import tsPlugin from "@typescript-eslint/eslint-plugin";
 import tsParser from "@typescript-eslint/parser";
+import { Linter } from "eslint";
 import prettierConfig from "eslint-config-prettier";
 import reactPlugin from "eslint-plugin-react";
 import reactHooksPlugin from "eslint-plugin-react-hooks";
+import Config = Linter.Config;
 
-export default [
+const config: Config[] = [
     js.configs.recommended,
     {
         files: ["**/*.{ts,tsx}"],
         plugins: {
+            // @ts-ignore
             "@typescript-eslint": tsPlugin,
         },
         languageOptions: {
             parser: tsParser,
+            globals: {
+                console: true,
+                process: true,
+                window: true,
+                setTimeout: true,
+                localStorage: true,
+                NodeJS:true,
+                clearInterval:true
+                // Add any other globals you need here
+            },
             parserOptions: {
                 ecmaVersion: "latest",
                 sourceType: "module",
@@ -48,3 +61,5 @@ export default [
     },
     prettierConfig,
 ];
+
+export default config;