+++ src/access-control/access-control-provider.tsx
@@ -5,13 +5,14 @@ import { AccessControlConfig, AuthUser } from "./type";
  * Type for the access control context
  * @template T - Type that extends AuthUser
  */
-type AccessControlContextType<T extends AuthUser = AuthUser> = AccessControlConfig<T> | null;
+type AccessControlContextType<T extends AuthUser = AuthUser> =
+    AccessControlConfig<T> | null;
 
 /**
  * React context for access control configuration
  * Using Record<string, any> as a workaround for TypeScript's limitation with generic context
  */
-const AccessControlContext = createContext<AccessControlContextType<AuthUser>>(null);
+const AccessControlContext = createContext<AccessControlContextType>(null);
 
 /**
  * Provider component for access control
@@ -44,8 +45,15 @@ export const AccessControlProvider = <T extends AuthUser>({
  * @throws Will throw an error if used outside an AccessControlProvider.
  * @returns {AccessControlConfig<T>} - The access control configuration.
  */
-export const useAccessControl = <T extends AuthUser>(): AccessControlConfig<T> => {
-    const context = useContext(AccessControlContext) as AccessControlContextType<T>;
-    if (!context) throw new Error("useAccessControl should be used inside AccessControlProvider");
+export const useAccessControl = <
+    T extends AuthUser,
+>(): AccessControlConfig<T> => {
+    const context = useContext(
+        AccessControlContext
+    ) as AccessControlContextType<T>;
+    if (!context)
+        throw new Error(
+            "useAccessControl should be used inside AccessControlProvider"
+        );
     return context;
 };